requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.74
langchain_openai==0.3.31
pydantic==2.11.7
python-dotenv==1.1.1
pypdf
uvicorn
chromadb
docx2txt
moviepy
whisper
